Remove deprecated method
URI.unescape has been depcrecated since Ruby 1.9. We now use CGI.unescape instead.

The behaviour of this method is a little different. For instance it unescapes '+' to space, so we need to encode all pluses before we begin.

It also doesn't raise an exception for invalid UTF-8 byte sequences so we can remove this protection.